requirements.txt
asgiref==3.8.1
Django==5.1.7
pillow==11.2.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
sqlparse==0.5.3
djangorestframework
celery
django-celery-beat